moved error mapper to controller space, separate from the seats api
