+++ controllers/seats.go
@@ -5,13 +5,12 @@ package controllers
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 
+	"github.com/RedHatInsights/entitlements-api-go/config"
 	"github.com/RedHatInsights/entitlements-api-go/logger"
 	v1 "github.com/openshift-online/ocm-sdk-go/accountsmgmt/v1"
-	ocmErrors "github.com/openshift-online/ocm-sdk-go/errors"
 	"github.com/sirupsen/logrus"
 
 	"github.com/RedHatInsights/entitlements-api-go/ams"
@@ -23,7 +22,6 @@ import (
 type SeatManagerApi struct {
 	ams 		ams.AMSInterface
 	bop    		bop.Bop
-	amsErrMapper	ams.AMSErrorMapper
 }
 
 const BASE_LINK_URL = "/api/entitlements/v1/seats"
@@ -34,51 +32,14 @@ func NewSeatManagerApi(amsClient ams.AMSInterface, bopClient bop.Bop, amsErrMapp
 	return &SeatManagerApi{
 		ams: amsClient,
 		bop: bopClient,
-		amsErrMapper: amsErrMapper,
 	}
 }
 
-// mapResponse will create a response based on the provided error
-// if err is of a more meaningful type than error, the response status will be set to what the error dictates
-// otherwise, the provided code will be used
-func mapResponse(s *SeatManagerApi, err error, httpStatusCode int) api.Error {
-	var amsError *ocmErrors.Error
-	if errors.As(err, &amsError) {
-		reason := s.amsErrMapper.MapErrorMessage(amsError)
-		return api.Error{
-			Error: 			toPtr(reason),
-			Code:  			toPtr(amsError.Code()),
-			Identifier: 	toPtr(amsError.ID()),
-			OperationId: 	toPtr(amsError.OperationID()),
-			Status: 		toPtr(amsError.Status()),
-		}
-	} 
-	
-	var clientError *ams.ClientError
-	if errors.As(err, &clientError) {
-		return api.Error{
-			Error: 	toPtr(clientError.Error()),
-			Status: toPtr(clientError.StatusCode),
-		}
-	} 
-
-	var userDetailErr *bop.UserDetailError
-	if errors.As(err, &userDetailErr) {
-		return api.Error{
-			Error: 	toPtr(userDetailErr.Error()),
-			Status: toPtr(userDetailErr.StatusCode),
-		}
-	}
-
-	return api.Error{
-		Error: toPtr(err.Error()),
-		Status: toPtr(httpStatusCode),
-	}
-}
+var errorMapper SeatsErrorMapper = NewErrorMapper(config.GetConfig())
 
 // doError will construct an api.Error reponse and write it to the response writer
-func doError(w http.ResponseWriter, s *SeatManagerApi, httpStatusCode int, err error, source string) {
-	response := mapResponse(s, err, httpStatusCode)
+func doError(w http.ResponseWriter, httpStatusCode int, err error, source string) {
+	response := errorMapper.MapResponse(err, httpStatusCode)
 
 	log := logger.Log.WithFields(logrus.Fields{"error": err, "status": httpStatusCode, "source": source})
 	if *response.Status == http.StatusInternalServerError {
@@ -96,38 +57,38 @@ func (s *SeatManagerApi) DeleteSeatsId(w http.ResponseWriter, r *http.Request, i
 	idObj := identity.Get(r.Context()).Identity
 
 	if !idObj.User.OrgAdmin {
-		doError(w, s, http.StatusForbidden, fmt.Errorf("Not allowed to delete subscription %s. User must be org admin", id), "")
+		doError(w, http.StatusForbidden, fmt.Errorf("Not allowed to delete subscription %s. User must be org admin", id), "")
 		return
 	}
 
 	subscription, err := s.ams.GetSubscription(id)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS GetSubscription")
+		doError(w, http.StatusInternalServerError, err, "AMS GetSubscription")
 		return
 	}
 
 	subOrgId, ok := subscription.GetOrganizationID()
 	if !ok {
-		doError(w, s, http.StatusInternalServerError, 
+		doError(w, http.StatusInternalServerError, 
 			fmt.Errorf("Subscription with id [%s] does not have a corresponding ams org id, cannot verify subscription org", id), "")
 		return
 	}
 
 	amsUserOrgId, err := s.ams.ConvertUserOrgId(idObj.Internal.OrgID)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS ConvertUserOrgId")
+		doError(w, http.StatusInternalServerError, err, "AMS ConvertUserOrgId")
 		return
 	}
 
 	if subOrgId != amsUserOrgId {
-		doError(w, s, http.StatusForbidden,
+		doError(w, http.StatusForbidden,
 			fmt.Errorf("Not allowed to delete subscription %s. Subscription org [%s] must match user ams org id [%s]}. User org [%s]",
 				id, subOrgId, amsUserOrgId, idObj.Internal.OrgID), "")
 		return
 	}
 
 	if err = s.ams.DeleteSubscription(id); err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS DeleteSubscription")
+		doError(w, http.StatusInternalServerError, err, "AMS DeleteSubscription")
 		return
 	}
 
@@ -159,12 +120,12 @@ func (s *SeatManagerApi) GetSeats(w http.ResponseWriter, r *http.Request, params
 	offset := int(*params.Offset)
 
 	if limit < 1 {
-		doError(w, s, http.StatusBadRequest, fmt.Errorf("limit must be > 0"), "")
+		doError(w, http.StatusBadRequest, fmt.Errorf("limit must be > 0"), "")
 		return
 	}
 
 	if offset < 0 {
-		doError(w, s, http.StatusBadRequest, fmt.Errorf("offset must be >= 0"), "")
+		doError(w, http.StatusBadRequest, fmt.Errorf("offset must be >= 0"), "")
 		return
 	}
 
@@ -172,13 +133,13 @@ func (s *SeatManagerApi) GetSeats(w http.ResponseWriter, r *http.Request, params
 
 	subs, err := s.ams.GetSubscriptions(idObj.Internal.OrgID, params, limit, page)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS GetSubscriptions")
+		doError(w, http.StatusInternalServerError, err, "AMS GetSubscriptions")
 		return
 	}
 
 	quotaCost, err := s.ams.GetQuotaCost(idObj.Internal.OrgID)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS GetQuotaCost")
+		doError(w, http.StatusInternalServerError, err, "AMS GetQuotaCost")
 		return
 	}
 
@@ -226,7 +187,7 @@ func (s *SeatManagerApi) GetSeats(w http.ResponseWriter, r *http.Request, params
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	if err = json.NewEncoder(w).Encode(resp); err != nil {
-		doError(w, s, http.StatusInternalServerError, fmt.Errorf("Unexpected error encoding response [%w]", err), "")
+		doError(w, http.StatusInternalServerError, fmt.Errorf("Unexpected error encoding response [%w]", err), "")
 		return
 	}
 
@@ -236,46 +197,46 @@ func (s *SeatManagerApi) PostSeats(w http.ResponseWriter, r *http.Request) {
 	idObj := identity.Get(r.Context()).Identity
 
 	if !idObj.User.OrgAdmin {
-		doError(w, s, http.StatusForbidden, fmt.Errorf("Not allowed to assign seats, must be an org admin."), "")
+		doError(w, http.StatusForbidden, fmt.Errorf("Not allowed to assign seats, must be an org admin."), "")
 		return
 	}
 
 	seat := new(api.SeatRequest)
 	defer r.Body.Close()
 	if err := json.NewDecoder(r.Body).Decode(seat); err != nil {
-		doError(w, s, http.StatusBadRequest, fmt.Errorf("PostSeats [%w]", err), "")
+		doError(w, http.StatusBadRequest, fmt.Errorf("PostSeats [%w]", err), "")
 		return
 	}
 
 	user, err := s.bop.GetUser(seat.AccountUsername)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "BOP GetUser")
+		doError(w, http.StatusInternalServerError, err, "BOP GetUser")
 		return
 	}
 
 	if user.OrgId != idObj.Internal.OrgID {
-		doError(w, s, http.StatusForbidden, fmt.Errorf("Not allowed to assign seats to users outside of Organization %s", idObj.Internal.OrgID), "")
+		doError(w, http.StatusForbidden, fmt.Errorf("Not allowed to assign seats to users outside of Organization %s", idObj.Internal.OrgID), "")
 		return
 	}
 
 	quotaCost, err := s.ams.GetQuotaCost(idObj.Internal.OrgID)
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS GetQuotaCost")
+		doError(w, http.StatusInternalServerError, err, "AMS GetQuotaCost")
 		return
 	}
 
 	resp, err := s.ams.QuotaAuthorization(seat.AccountUsername, quotaCost.Version())
 	if err != nil {
-		doError(w, s, http.StatusInternalServerError, err, "AMS QuotaAuthorization")
+		doError(w, http.StatusInternalServerError, err, "AMS QuotaAuthorization")
 		return
 	}
 
 	if !resp.Allowed() {
 		if len(resp.ExcessResources()) > 0 {
-			doError(w, s, http.StatusConflict, fmt.Errorf("Assignment request was denied due to excessive resource requests"), "")
+			doError(w, http.StatusConflict, fmt.Errorf("Assignment request was denied due to excessive resource requests"), "")
 			return
 		}
-		doError(w, s, http.StatusForbidden, fmt.Errorf("Assignment request was denied"), "")
+		doError(w, http.StatusForbidden, fmt.Errorf("Assignment request was denied"), "")
 		return
 	}
 
@@ -289,7 +250,7 @@ func (s *SeatManagerApi) PostSeats(w http.ResponseWriter, r *http.Request) {
 		SubscriptionId:  &subId,
 		AccountUsername: &userName,
 	}); err != nil {
-		doError(w, s, http.StatusInternalServerError, fmt.Errorf("Unexpected error encoding response [%w]", err), "")
+		doError(w, http.StatusInternalServerError, fmt.Errorf("Unexpected error encoding response [%w]", err), "")
 		return
 	}
 }
